refactor: migrate headers and rename namespace
