+++ Makefile
@@ -1,29 +1,29 @@
-FLAGS = -Wall -Wextra
+FLAGS = -Wall -Wextra -I./include
 
 ifeq ($(OS),Windows_NT)
 GTEST_LIBS = -L/mingw64/lib -lgtest -lbcrypt
 PLATFORM_LIBS = -lbcrypt
 else
-GTEST_LIBS = -pthread /usr/lib/libgtest.a
+GTEST_LIBS = -pthread /usr/lib/x86_64-linux-gnu/libgtest.a /usr/lib/x86_64-linux-gnu/libgtest_main.a
 PLATFORM_LIBS =
 endif
 
 build_all: clean build_test build_debug build_profile build_release build_speed_test
 
 build_test:
-	docker-compose exec aes g++ $(FLAGS) -g -pthread ./src/AES.cpp ./src/AESUtils.cpp ./tests/tests.cpp /usr/lib/libgtest.a -o bin/test
+	docker-compose exec aes g++ $(FLAGS) -g -pthread ./src/aes.cpp ./src/aes_utils.cpp ./tests/tests.cpp /usr/lib/libgtest.a -o bin/test
 
 build_debug:
-	docker-compose exec aes g++ $(FLAGS) -g ./src/AES.cpp ./src/AESUtils.cpp ./dev/main.cpp -o bin/debug
+	docker-compose exec aes g++ $(FLAGS) -g ./src/aes.cpp ./src/aes_utils.cpp ./dev/main.cpp -o bin/debug
 
 build_profile:
-	docker-compose exec aes g++ $(FLAGS) -pg ./src/AES.cpp ./src/AESUtils.cpp ./dev/main.cpp -o bin/profile
+	docker-compose exec aes g++ $(FLAGS) -pg ./src/aes.cpp ./src/aes_utils.cpp ./dev/main.cpp -o bin/profile
 
 build_speed_test:
-	docker-compose exec aes g++ $(FLAGS) -O2 ./src/AES.cpp ./src/AESUtils.cpp ./speedtest/main.cpp -o bin/speedtest
+	docker-compose exec aes g++ $(FLAGS) -O2 ./src/aes.cpp ./src/aes_utils.cpp ./speedtest/main.cpp -o bin/speedtest
 
 build_release:
-	docker-compose exec aes g++ $(FLAGS) -O2 ./src/AES.cpp ./src/AESUtils.cpp ./dev/main.cpp -o bin/release
+	docker-compose exec aes g++ $(FLAGS) -O2 ./src/aes.cpp ./src/aes_utils.cpp ./dev/main.cpp -o bin/release
 
 
 
@@ -43,15 +43,16 @@ speed_test:
 	docker-compose exec aes bin/speedtest
 
 style_fix:
-	docker-compose exec aes bash -c "clang-format -i src/*.cpp src/*.h tests/*.cpp"
+	docker-compose exec aes bash -c "clang-format -i src/*.cpp include/aescpp/*.hpp tests/*.cpp"
+
 
 clean:
 	docker-compose exec aes rm -rf bin/*
 
 
 workflow_build_test:
-	g++ $(FLAGS) -g ./src/AES.cpp ./src/AESUtils.cpp ./tests/tests.cpp $(GTEST_LIBS) -o bin/test
+	g++ $(FLAGS) -g ./src/aes.cpp ./src/aes_utils.cpp ./tests/tests.cpp $(GTEST_LIBS) -o bin/test
 
 workflow_build_speed_test:
-	g++ $(FLAGS) -O2 ./src/AES.cpp ./src/AESUtils.cpp ./speedtest/main.cpp $(PLATFORM_LIBS) -o bin/speedtest
+	g++ $(FLAGS) -O2 ./src/aes.cpp ./src/aes_utils.cpp ./speedtest/main.cpp $(PLATFORM_LIBS) -o bin/speedtest
