+++ README.md
@@ -14,6 +14,8 @@ ECB, CBC, CFB, CTR and GCM modes are implemented. GCM additionally produces an a
 ## Vector Overloads
 All encryption and decryption methods have overloads that accept `std::vector<unsigned char>` in addition to raw pointer APIs:
 ```c++
+#include <aescpp/aes.hpp>
+
 std::vector<unsigned char> plainVec = { 0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff };
 std::vector<unsigned char> keyVec   = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f };
 unsigned char iv[]                  = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f };
@@ -26,6 +28,8 @@ auto cipherVec = aesVec.EncryptCBC(plainVec, keyVec, iv);
 
 ### Encryption/Decryption
 ```c++
+#include <aescpp/aes.hpp>
+
 unsigned char plain[] = { 0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff };
 unsigned char key[]   = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f };
 unsigned char iv[]    = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f };
@@ -38,6 +42,8 @@ auto restored = aes.DecryptCBC(cipher.get(), sizeof(plain), key, iv);
 ### GCM Tagging
 GCM mode requires a 12-byte (96-bit) IV.
 ```c++
+#include <aescpp/aes.hpp>
+
 unsigned char gcm_iv[12] = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05,
                              0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b };
 unsigned char tag[16];
