+++ dev/main.cpp
@@ -1,6 +1,5 @@
+#include <aescpp/aes.hpp>
 #include <iostream>
-
-#include "../src/AES.h"
 
 int main() {
   std::cout << "Aes dev" << std::endl;
