+++ src/aes.cpp
@@ -1,5 +1,4 @@
-#include "AES.h"
-
+#include <aescpp/aes.hpp>
 #include <algorithm>
 #include <cstddef>
 #include <cstdint>
@@ -265,7 +264,7 @@ unsigned char *AES::EncryptGCM(const unsigned char in[], size_t inLen,
   // Шифрование данных в режиме CTR
   unsigned char ctr[16] = {0};
   memcpy(ctr, iv, 12);  // IV занимает 12 байт
-  ctr[15] = 1;  // Установить начальное значение счетчика
+  ctr[15] = 1;          // Установить начальное значение счетчика
   unsigned char encryptedCtr[16] = {0};
 
   // GHASH для AAD без промежуточных буферов
